Update to egui 0.24.1. Replace `bevy_demo` with demo using `eframe`.
Also removes the custom style of the demo. Custom style can always be
applied downstream.

No major changes required for the `egui` version update.

The switch to eframe keeps the example a little simpler and removes the
arguably unrelated bevy dependencies. Hopefully makes it a little easier
to maintain as the `egui` and `eframe` are both maintained by egui-rs.