refactor: Graph show calls a user fn rather than return a session type
Previously `Graph::show` returned a session type on which the nodes and
edges could be instantiated. This has been changed to take a user
function as an argument that can be called to add content.

This is a bit rough for now, but better matches the `Scene::show`
implementation that we'll use in the graph for enabling pan+zoom.