fix: Swap out node Window for custom Frame
This ensures that node widgets placed have the same `Layer` as the
underlying movable `Scene`. `Window` and `Area` widgets have a separate
layer from the `Ui` with which they're instantiated, which means
previously the node wouldn't pan or zoom with the `Scene`'s `Ui`.

Still need to work out how to handle the `Response` sensing for the
`Frame` and child widgets. For now, to test that clicking/dragging is
possible, we just have a `.interact` call on the response returned from
the `Frame`.